requirements.txt
asgiref==3.5.2
Django==4.0.5
django-colorfield==0.7.1
django-multiselectfield==0.1.12
django-phonenumber-field==6.3.0
django-widget-tweaks==1.4.12
phonenumbers==8.12.50
Pillow==9.2.0
sqlparse==0.4.2
tzdata==2022.1
